Move utils, builder under master
